+++ packages/styles/src/packages/tooltip/index.ts
@@ -1,0 +1,5 @@
+import * as style from "./tooltip.style.css";
+
+export const tooltip = {
+  style,
+};