+++ packages/styles/CHANGELOG.md
@@ -2,6 +2,12 @@
 
 Nimbus Styles deprive all styles needed to build nimbus components.
 
+## 2022-10-01
+
+### 🎉 New features
+
+- Added new style pack for tooltip component. ([#34](https://github.com/TiendaNube/nimbus-design-system/pull/34) by [@juniorconquista](https://github.com/juniorconquista))
+
 ## 2022-09-28
 
 ### 🎉 New features